Simon Game Progress
Restarted Development. Tested the finished product and the expectation was different to the progress I'd made. Have implemented and unit tested the following successfully: gameover functionality, user button press visual functionality (minus audio queues) and to check if a particular button press is correct in accordance to the generated list (if empty, defaults to a game over sequence).